Refactor Challenge and Solution types
Challenge and Solution types were previously defined as tuples. This
made type signatures busy and hard to read. Instead, they're now
defined as record types. The functions for extracting values have
remained for the time being, and will be removed in favour of direct
field access in the future.

Future refactorings will make creation and mutation cleaner.  # On
branch master